Display meaningful error when trying to delete language that is already assigned to a teacher

+++ src/TeacherSuite.Application/ProgrammingLanguages/Commands/Delete/DeleteProgrammingLanguage.cs
@@ -1,3 +1,4 @@
+using TeacherSuite.Application.Common;
 using TeacherSuite.Application.Common.Interfaces;
 
 namespace TeacherSuite.Application.ProgrammingLanguages.Commands.Delete;
@@ -11,6 +12,14 @@ public class DeleteProgrammingLanguageHandler(IApplicationDbContext context) : I
         var entity = await context.ProgrammingLanguages.FindAsync(new object[] { request.Id }, cancellationToken);
 
         Guard.Against.NotFound(request.Id, entity);
+
+        var isAssigned = await context.TeacherProgrammingLanguages
+            .AnyAsync(tpl => tpl.ProgrammingLanguageId == request.Id, cancellationToken);
+
+        if (isAssigned)
+        {
+            throw new ConflictException("The programming language is assigned to a teacher and cannot be deleted.");
+        }
 
         context.ProgrammingLanguages.Remove(entity);
         await context.SaveChangesAsync(cancellationToken);
